Replaced previous implementation to use a generic class to create checkbox options
Changed editor templates so they can be easily invoked using editor for 2nd arg specifying template name (built into checkbox class already)
