+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForDataBreachImpactManagement.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.DataModificationOptions
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.DataBreachImpactManagement>
 <div class="form-group">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
