+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForDataBreachPlan.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.EnvironmentReconfigurationOptions
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.DataBreachPlan>
 <div class="form-group">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
