+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForDataModification.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.DataBreachPlanOptions
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.DataModification>
 <div class="form-group">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
