+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForEnvironmentReconfigurations.cshtml
@@ -1,5 +1,5 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.DatasetProperties
-<div class="form-group form-check">
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.EnvironmentReconfigurations>
+<div class="form-group">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
         <label class="control-label">
