+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForProperties.cshtml
@@ -1,5 +1,5 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.SpecificationOutcomes
-<div class="form-group">
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.Properties>
+<div class="form-group form-check">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
         <label class="control-label">
