+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForSpecifications.cshtml
@@ -1,4 +1,5 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.StakeholderAssuranceDataProtection
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.Specifications>
+
 <div class="form-group">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
