+++ solution/WebApplication/WebApplication/Views/Shared/EditorTemplates/EditorForStakeholderAssuranceProtection.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebApplication.Models.Customisations.WizardReference.DataBreachImpactManagementOptions
+﻿@model WebApplication.Models.Customisations.WizardReference.CheckBoxOptionsFor<WebApplication.Models.Customisations.WizardReference.StakeholderAssuranceProtection>
 <div class="form-group">
     @for (int i = 0; i < Model.Options.Count; i++)
     {
